+++ inc/taskjobview.class.php
@@ -31,6 +31,7 @@
  */
 
 use Glpi\DBAL\QueryExpression;
+use Glpi\Exception\Http\BadRequestHttpException;
 use Glpi\Exception\Http\NotFoundHttpException;
 
 use function Safe\preg_match;
@@ -818,7 +819,7 @@ class PluginGlpiinventoryTaskjobView extends PluginGlpiinventoryCommonView
         if (isset($postvars['definition_add'])) {
             // * Add a definition
             Session::checkRight(PluginGlpiinventoryTask::$rightname, UPDATE);
-            $mytaskjob->getFromDB($postvars['id']);
+            $mytaskjob->check($postvars['id'], UPDATE);
             $a_listdef = importArrayFromDB($mytaskjob->fields['definition']);
             $add = 1;
             foreach ($a_listdef as $dataDB) {
@@ -844,7 +845,7 @@ class PluginGlpiinventoryTaskjobView extends PluginGlpiinventoryCommonView
         } elseif (isset($postvars['action_add'])) {
             // * Add an action
             Session::checkRight(PluginGlpiinventoryTask::$rightname, UPDATE);
-            $mytaskjob->getFromDB($postvars['id']);
+            $mytaskjob->check($postvars['id'], UPDATE);
             $a_listact = importArrayFromDB($mytaskjob->fields['action']);
             $add = 1;
             foreach ($a_listact as $dataDB) {
@@ -870,7 +871,7 @@ class PluginGlpiinventoryTaskjobView extends PluginGlpiinventoryCommonView
         } elseif (isset($postvars['definition_delete'])) {
             // * Delete definition
             Session::checkRight(PluginGlpiinventoryTask::$rightname, UPDATE);
-            $mytaskjob->getFromDB($postvars['id']);
+            $mytaskjob->check($postvars['id'], UPDATE);
             $a_listdef = importArrayFromDB($mytaskjob->fields['definition']);
 
             foreach ($postvars['definition_to_delete'] as $itemdelete) {
@@ -889,7 +890,7 @@ class PluginGlpiinventoryTaskjobView extends PluginGlpiinventoryCommonView
         } elseif (isset($postvars['action_delete'])) {
             // * Delete action
             Session::checkRight(PluginGlpiinventoryTask::$rightname, UPDATE);
-            $mytaskjob->getFromDB($postvars['id']);
+            $mytaskjob->check($postvars['id'], UPDATE);
             $a_listact = importArrayFromDB($mytaskjob->fields['action']);
 
             foreach ($postvars['action_to_delete'] as $itemdelete) {
@@ -912,6 +913,7 @@ class PluginGlpiinventoryTaskjobView extends PluginGlpiinventoryCommonView
             $pfTaskjob = new PluginGlpiinventoryTaskjob();
             $_SESSION["plugin_glpiinventory_forcerun"] = [];
             foreach ($postvars['taskjobstoforcerun'] as $taskjobs_id) {
+                $pfTaskjob->check($taskjobs_id, UPDATE);
                 $pfTask->getFromDB($pfTaskjob->fields['plugin_glpiinventory_tasks_id']);
                 $pfTask->forceRunning();
             }
@@ -922,13 +924,17 @@ class PluginGlpiinventoryTaskjobView extends PluginGlpiinventoryCommonView
                 $postvars['entities_id'] ??= $_SESSION['glpidefault_entity'] ?? 0;
                 // Get entity of task
                 $pfTask = new PluginGlpiinventoryTask();
-                $pfTask->getFromDB($postvars['plugin_glpiinventory_tasks_id']);
+                $pfTask->check($postvars['plugin_glpiinventory_tasks_id'], UPDATE);
                 $entities_list = getSonsOf('glpi_entities', $pfTask->fields['entities_id']);
                 if (!in_array($postvars['entities_id'], $entities_list)) {
                     $postvars['entities_id'] = $pfTask->fields['entities_id'];
                 }
                 $jobs_id = $this->add($postvars);
             } else {
+                $this->check($postvars['id'], UPDATE);
+                if (isset($postvars['plugin_glpiinventory_tasks_id'])) {
+                    (new PluginGlpiinventoryTask())->check($postvars['plugin_glpiinventory_tasks_id'], UPDATE);
+                }
                 if (isset($postvars['method_id'])) {
                     $postvars['method']  = $postvars['method_id'];
                 }
@@ -971,10 +977,16 @@ class PluginGlpiinventoryTaskjobView extends PluginGlpiinventoryCommonView
         } elseif (isset($postvars["delete"])) {
             // * delete taskjob
             Session::checkRight(PluginGlpiinventoryTask::$rightname, PURGE);
+            $this->check($postvars['id'], PURGE);
 
             $this->delete($postvars);
         } elseif (isset($postvars['itemaddaction'])) {
             Session::checkRight(PluginGlpiinventoryTask::$rightname, CREATE);
+            $target = getItemForItemtype($postvars['itemtype'] ?? '');
+            if (!$target) {
+                throw new BadRequestHttpException();
+            }
+            $target->check((int) ($postvars['items_id'] ?? 0), READ);
             $array                     = explode("||", $postvars['methodaction']);
             $module                    = $array[0];
             $method                    = $array[1];
@@ -1008,10 +1020,14 @@ class PluginGlpiinventoryTaskjobView extends PluginGlpiinventoryCommonView
             $mytask->update($mytask->fields);
             // force running this job (?)
         } elseif (isset($postvars['forceend'])) {
+            Session::checkRight(PluginGlpiinventoryTask::$rightname, UPDATE);
             $mytaskjobstate = new PluginGlpiinventoryTaskjobstate();
             $pfTaskjob = new PluginGlpiinventoryTaskjob();
-            $mytaskjobstate->getFromDB($postvars['taskjobstates_id']);
+            if (!$mytaskjobstate->getFromDB($postvars['taskjobstates_id'])) {
+                throw new NotFoundHttpException();
+            }
             $jobstate = $mytaskjobstate->fields;
+            $pfTaskjob->check($jobstate['plugin_glpiinventory_taskjobs_id'], UPDATE);
             $a_taskjobstates = $mytaskjobstate->find(['uniqid' => $mytaskjobstate->fields['uniqid']]);
             foreach ($a_taskjobstates as $data) {
                 if ($data['state'] != PluginGlpiinventoryTaskjobstate::FINISHED) {
@@ -1028,7 +1044,9 @@ class PluginGlpiinventoryTaskjobView extends PluginGlpiinventoryCommonView
             $pfTaskjob->getFromDB($jobstate['plugin_glpiinventory_taskjobs_id']);
             $pfTaskjob->reinitializeTaskjobs($pfTaskjob->fields['plugin_glpiinventory_tasks_id']);
         } elseif (isset($postvars['delete_taskjobs'])) {
+            Session::checkRight(PluginGlpiinventoryTask::$rightname, PURGE);
             foreach ($postvars['taskjobs'] as $taskjob_id) {
+                $this->check($taskjob_id, PURGE);
                 $input = ['id' => $taskjob_id];
                 $this->delete($input, true);
             }